+++ Tests/IntegrationTests.Shared/NotificationTests.cs
@@ -1,0 +1,37 @@
+﻿using System;
+using NUnit.Framework;
+using Realms;
+using System.Threading;
+using System.IO;
+
+namespace IntegrationTests.Shared
+{
+    [TestFixture]
+    public class NotificationTests
+    {
+        private string _databasePath;
+        private Realm _realm;
+
+        [SetUp]
+        public void Setup()
+        {
+            _databasePath = Path.GetTempFileName();
+            _realm = Realm.GetInstance(_databasePath);
+        }
+
+        [Test]
+        public void ShouldTriggerRealmChangedEvent() 
+        {
+            // Arrange
+            var wasNotified = false;
+            _realm.RealmChanged += (sender, e) => { wasNotified = true; };
+
+            // Act
+            _realm.Write(() => _realm.CreateObject<Person>());
+
+            // Assert
+            Assert.That(wasNotified, "RealmChanged notification was not triggered");
+        }
+    }
+}
+