+++ Tests/IntegrationTests.Shared/SimpleLINQtests.cs
@@ -13,10 +13,8 @@ namespace IntegrationTests
 {
     class SimpleLINQtests : PeopleTestsBase
     {
-
-
-        [SetUp]
-        public new void Setup()
+        // see comment on base method why this isn't decorated with [SetUp]
+        public override void Setup()
         {
             base.Setup();
             MakeThreePeople();
@@ -38,7 +36,7 @@ namespace IntegrationTests
             Assert.That(s2[0].Email, Is.EqualTo("john@doe.com"));
             Assert.That(s2[1].Email, Is.EqualTo("peter@jameson.com"));
 
-            var s3 = _realm.All<Person>().Where(p => p.Email != "").ToList();
+            var s3 = _realm.All<Person>().Where(p => p.Email != "");
             Assert.That(s3.Count(), Is.EqualTo(3));
         }
 
@@ -48,6 +46,9 @@ namespace IntegrationTests
         {
             var countSimpleNot = _realm.All<Person>().Where(p => !p.IsInteresting).Count();
             Assert.That(countSimpleNot, Is.EqualTo(1));
+
+            var countSimpleNot2 = _realm.All<Person>().Count(p => !p.IsInteresting);
+            Assert.That(countSimpleNot2, Is.EqualTo(1));
 
             var countNotEqual = _realm.All<Person>().Where(p => !(p.Score == 42.42f)).Count();
             Assert.That(countNotEqual, Is.EqualTo(2));
@@ -71,6 +72,9 @@ namespace IntegrationTests
 
             var c3 = _realm.All<Person>().Where(p => p.FirstName=="John").Count();
             Assert.That(c3, Is.EqualTo(2));
+
+            var c4 = _realm.All<Person>().Count(p => p.FirstName=="John");
+            Assert.That(c4, Is.EqualTo(2));
         }