+++ Tests/IntegrationTests.Shared/TestHelpers.cs
@@ -31,12 +31,24 @@ namespace IntegrationTests
 
         public static void CopyBundledDatabaseToDocuments(string realmName, string destPath=null, bool overwrite=true)
         {
+            destPath = RealmConfiguration.PathToRealm(destPath);  // any relative subdir or filename works
+
+            #if __ANDROID__
+            using (var asset = Android.App.Application.Context.Assets.Open(realmName))
+            using (var destination = File.OpenWrite(destPath))
+            {
+                asset.CopyTo(destination);
+            }
+
+            return;
+            #endif
+
             string sourceDir = "";
             #if __IOS__
             sourceDir = NSBundle.MainBundle.BundlePath;
             #endif
-            //TODO add cases for Android and Windows setting sourcedir for bundled files
-            destPath = RealmConfiguration.PathToRealm(destPath);  // any relative subdir or filename works
+            //TODO add cases for Windows setting sourcedir for bundled files
+
             File.Copy(Path.Combine(sourceDir, realmName), destPath, overwrite);
         }