+++ wrappers/src/error_handling.cpp
@@ -6,6 +6,7 @@
 #include <sstream>
 #include <cassert>
 #include "object-store/src/shared_realm.hpp"
+#include "object-store/src/object_store.hpp"
 #include "wrapper_exceptions.hpp"
 #include "realm_export_decls.hpp"
 #include "error_handling.hpp"